+++ src/main/java/shop/mtcoding/boardproject/user/UserController.java
@@ -2,38 +2,16 @@ package shop.mtcoding.boardproject.user;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-
 import shop.mtcoding.boardproject._core.error.ex.MyException;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import shop.mtcoding.boardproject._core.util.ApiUtil;
-import shop.mtcoding.boardproject.apply.Apply;
-import shop.mtcoding.boardproject.apply.ApplyService;
 import shop.mtcoding.boardproject.comp.CompRequest;
-import shop.mtcoding.boardproject.comp.CompService;
-import shop.mtcoding.boardproject.posting.Posting;
-import shop.mtcoding.boardproject.posting.PostingRequest.CompInfoDTO;
-import shop.mtcoding.boardproject.recommend.Recommend;
-import shop.mtcoding.boardproject.recommend.RecommendService;
-import shop.mtcoding.boardproject.resume.Resume;
-import shop.mtcoding.boardproject.resume.ResumeService;
-import shop.mtcoding.boardproject.skill.PostingSkill;
-import shop.mtcoding.boardproject.skill.Skill;
-import shop.mtcoding.boardproject.skill.SkillService;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Controller
 public class UserController {
@@ -43,71 +21,8 @@ public class UserController {
     private UserService userService;
 
     @Autowired
-    private ResumeService resumeService;
-
-    @Autowired
-    private ApplyService applyService;
-
-    @Autowired
-    private CompService compService;
-
-    @Autowired
-    private RecommendService recommendService;
-
-    @Autowired
-    private SkillService skillService;
-
-    @Autowired
     private HttpSession session;
 
-    // 17_개인기업추천 화면
-    @GetMapping("/user/recommend/form")
-    public String userRecommendForm(
-            @RequestParam(defaultValue = "all") List<String> skillList,
-            @RequestParam(defaultValue = "all") String position,
-            HttpServletRequest request) {
-
-        User sessionUser = (User) session.getAttribute("sessionUser");
-        if (sessionUser == null) {
-            return "redirect:/user/loginForm";
-        }
-
-        List<Skill> sl = skillService.스킬이름전부();
-
-        request.setAttribute("skillList", sl);
-
-        request.setAttribute("position", position);
-        try {
-            ObjectMapper objectMapper = new ObjectMapper();
-            String json = objectMapper.writeValueAsString(skillList);
-            // System.out.println("테스트"+json);
-            request.setAttribute("json", json);
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        }
-
-        List<Posting> compList = userService.기업추천검색(skillList, position);
-        System.out.println("검색1: " + compList);
-
-        // CompInfoDTO 리스트 생성
-        List<CompInfoDTO> compInfoList = new ArrayList<>();
-        for (Posting posting : compList) {
-            CompInfoDTO compInfoDTO = new CompInfoDTO();
-            compInfoDTO.setId(posting.getId());
-            compInfoDTO.setCompname(posting.getUser().getCompname());
-            compInfoDTO.setTitle(posting.getTitle());
-            compInfoDTO.setPosition(posting.getPosition());
-            // 공고에 해당하는 스킬 정보 가져오기
-            List<PostingSkill> postingSkills = skillService.공고별스킬조회(posting.getId());
-            compInfoDTO.setPostingSkills(postingSkills);
-            compInfoList.add(compInfoDTO);
-        }
-
-        // compInfoList를 컨트롤러에서 뷰로 전달
-        request.setAttribute("compInfoList", compInfoList);
-
-        return "user/recommend/form";
-    }
 
     // 12번 수정하기 버튼 POST
     @PostMapping("/user/update")
@@ -135,44 +50,6 @@ public class UserController {
         return "user/updateForm";
     }
 
-    // 11번 지원하기 버튼 POST
-    @PostMapping("/user/apply")
-    public String userApply(@RequestParam("selectedResume") Integer selectedResumeId,
-            @RequestParam("postingId") Integer postingId) {
-        User sessionUser = (User) session.getAttribute("sessionUser");
-
-        if (sessionUser == null) {
-            return "redirect:/user/loginForm";
-        }
-
-        Posting posting = compService.공고찾기(postingId);
-        Resume selectedResume = resumeService.이력서찾기(selectedResumeId); // 선택한 이력서를 ID로 조회
-        if (selectedResume.getDisclosure() == false) {
-            throw new MyException("비공개 이력서로 지원 못함");
-        }
-        Apply apply = new Apply();
-        apply.setUser(sessionUser);
-        apply.setResume(selectedResume);
-        apply.setPosting(posting); // Apply 엔티티에 공고 설정
-        applyService.지원(apply); // Apply 엔티티 저장
-        return "redirect:/";
-    }
-
-    // 개인_지원하기 페이지
-    @GetMapping("/user/apply/form/{postingId}")
-    public String userApplyForm(Model model, @PathVariable("postingId") Integer postingId) {
-        Posting posting = compService.공고찾기(postingId);
-
-        User sessionUser = (User) session.getAttribute("sessionUser");
-        Integer user = sessionUser.getId();
-        List<Resume> resumes = resumeService.이력서목록(user);
-
-        model.addAttribute("resumes", resumes); // 이력서 목록 모델에 추가
-        model.addAttribute("user", user); // 유저 아이디를 모델에 추가
-        model.addAttribute("posting", posting); // 공고를 모델에 추가
-
-        return "user/apply/form";
-    }
 
     // 5_로그인 화면
     @GetMapping("/user/loginForm")
@@ -184,6 +61,7 @@ public class UserController {
 
         return "user/loginForm"; // view 파일 호출 user/loginForm 파일 호출
     }
+
 
     // 로그인
     @PostMapping("/login")
@@ -250,20 +128,20 @@ public class UserController {
         return "redirect:/";
     }
 
-    // 3_개인회원가입 화면
+    // 개인회원가입 화면
     @GetMapping("/user/joinForm")
     public String userJoinForm() {
         return "user/joinForm"; // view 파일 호출 user/loginForm 파일 호출
     }
 
-    // 2번 회원가입 버튼 POST
+    // 회원가입 버튼 POST
     @PostMapping("/join")
     public String join(UserRequest.JoinDTO joinDTO) {
         userService.회원가입(joinDTO);
         return "user/loginForm";
     }
 
-    // 2_회원가입유형선택 화면
+    // 회원가입유형선택 화면
     @GetMapping("/selectJoinForm")
     public String selectJoinForm() {
         return "user/selectJoinForm";
@@ -280,56 +158,6 @@ public class UserController {
         return new ApiUtil<String>(true, "이메일을 사용 할 수 있습니다.");
     }
 
-    // 개인 이력서를 보고 기업에서 입사제안 하는 페이지
-    @GetMapping("/user/resume/{resumeId}/offer/list")
-    public String offerListUser(@PathVariable Integer resumeId, HttpServletRequest request) {
-        // 세션을 찾는다.
-        User sessionAllUser = (User) session.getAttribute("sessionAllUser");
-        // 이력서를 찾는다.
-        Resume resume = resumeService.이력서찾기(resumeId, sessionAllUser.getId());
-        // 세션id와 이력서의 유저id가 같은지 비교한다.
-        if (sessionAllUser.getId() != resume.getUser().getId()) {
-            throw new MyException("내 이력서가 아닙니다.");
-        }
-        // id가 동일하다면 request에 이력서를 담는다.
-        request.setAttribute("resume", resume);
-        // 리스트에 찾은 이력서를 넣는다.
-        List<Recommend> recommendList = recommendService.이력서받은오퍼찾기(resumeId);
-        // request에 추천리스트를 담는다.
-        request.setAttribute("recommendList", recommendList);
-        // 입사 제안 보기 페이지로 이동
-        return "user/offer/list";
-    }
-
-    // 입사제안에 대한 거절 POST
-    @PostMapping("/user/resume/offer/{recommendId}/fail")
-    public String offerFail(@PathVariable Integer recommendId) {
-        // 세션을 찾는다.
-        User sessionAllUser = (User) session.getAttribute("sessionAllUser");
-        // 세션이 없다면 로그인 화면 이동시킨다.
-        if (sessionAllUser == null) {
-            return "redirect:/user/loginForm";
-        }
-        // 세션의id와 추천id를 비교해서 거절
-        Integer resumeId = recommendService.오퍼거절(recommendId, sessionAllUser.getId());
-        // 입사제안 화면으로 리다이렉트
-        return "redirect:/user/resume/" + resumeId + "/offer/list";
-    }
-
-    // 입사제안에 대한 수락 POST
-    @PostMapping("/user/resume/offer/{recommendId}/pass")
-    public String offerPass(@PathVariable Integer recommendId) {
-        // 세션을 찾는다.
-        User sessionAllUser = (User) session.getAttribute("sessionAllUser");
-        // 세션이 없다면 로그인 화면 이동시킨다.
-        if (sessionAllUser == null) {
-            return "redirect:/user/loginForm";
-        }
-        // 세션의id와 추천id를 비교해서 수락
-        Integer resumeId = recommendService.오퍼수락(recommendId, sessionAllUser.getId());
-        // 입사제안 화면으로 리다이렉트
-        return "redirect:/user/resume/" + resumeId + "/offer/list";
-    }
 
     // // 기업추천 검색 POST
     // @PostMapping("/api/user/recommend")
